fix(product): rename inference_service more appropriately.
